+++ src/screens/ProfileScreen/ProfileScreen.tsx
@@ -43,7 +43,7 @@ import Text from '../../components/Text';
 const ProfileScreen: React.FC<any> = ({navigation, route}) => {
   const [profilestate, dispatch] = useReducer(profileReducer, initProfileState);
   const {
-    authContext: {getProfileAuth},
+    authContext: {getProfileAuth}
   } = useAuth();
   const backbotton = !route.params ? true : route.params.navbar;
   const windowWidth = Dimensions.get('screen').width;
@@ -382,6 +382,19 @@ const ProfileScreen: React.FC<any> = ({navigation, route}) => {
               </TouchableOpacity>
             )}
           </View>
+          {profilestate.status === 'REJECTED' ?
+          <View style={styles.commentBg}>
+          <Text style={styles.commentFront}>หมายเหตุ : บัตรประชาชนไม่ชัดเจน/ไม่ถูกต้อง 
+            กรุณาติดต่อเจ้าหน้าที่ เพื่อดำเนินการแก้ไข โทร. 02-233-9000</Text>
+        </View>: profilestate.status === 'INACTIVE' ?
+        
+        <View style={styles.commentBg}>
+        <Text style={styles.commentFront}>หมายเหตุ : หากต้องการเปิดใช้งานบัญชี 
+          กรุณาติดต่อเจ้าหน้าที่ โทร. 02-233-9000</Text>
+      </View>: <></>
+          
+          }
+          
           <View
             style={{
               paddingVertical: normalize(15),
@@ -1104,4 +1117,17 @@ const styles = StyleSheet.create({
     color: '#2EC66E',
     paddingEnd: normalize(8),
   },
+  commentBg: {
+    backgroundColor:'#FFF7F4',
+    borderColor:'#FEE9E1',
+    borderWidth:1,
+    paddingHorizontal:10,
+    paddingVertical:12
+  },
+  commentFront:{
+    fontFamily: font.medium,
+    fontSize: normalize(14),
+    color:'#242D35'
+  }
+
 });